requirements.txt
altgraph==0.17
certifi==2020.11.8
chardet==3.0.4
future==0.18.2
idna==2.10
pefile==2019.4.18
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyodbc==4.0.30
python-dateutil==2.8.1
pywin32==300
pywin32-ctypes==0.2.0
requests==2.25.0
six==1.15.0
tqdm==4.54.0
urllib3==1.26.2
WMI==1.5.1
